backend/requirements.txt
fastapi
requests
pandas
numpy
fpdf
webdav3
kubernetes
pyyaml
stripe
stable-baselines3
